+++ src/screens/HomeScreen.jsx
@@ -1,27 +1,31 @@
 import React from 'react';
-import { View, Text } from 'react-native';
+import { View } from 'react-native';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import NavBar from "../components/NavBar";
-import MyOwnButton from "../components/MyOwnButton"
-import StylesHomeScreen from "../styles/stylesHomeScreen"
-import Search from "./Search";
-import BuyCart from './BuyCart';
-import LoginUser from './LoginUser';
-import RegisterUser from './RegisterUser';
+import Search from '../components/Search';
+import Category from './Category';
+import Shopping from './Shopping';
+import Offer from './Offer';
+import Profile from './Profile';
+import Support from './Support';
+import Product from '../components/HomeContent';
+import Favorite from './Favorite';
 import DrawerNavigation from '../components/DrawerNavigation';
 
 const Drawer = createDrawerNavigator();
 
-const HomeContent = () => {
+const HomeContent = ({navigation}) => {
   return (
-    <View style={StylesHomeScreen.container}>
+    <View>
       <NavBar />
-      <View style={StylesHomeScreen.contentContainer}>
-        <Text style={StylesHomeScreen.text}>Contenido Principal de la Home Screen</Text>
+      <View>
+       <Product navigation={navigation}/>
       </View>
     </View>
   );
 };
+
+
 
 const HomeScreen = () => {
   return (
@@ -32,14 +36,17 @@ const HomeScreen = () => {
       }}
     >
       <Drawer.Screen name="Inicio" component={HomeContent} />
-      <Drawer.Screen name="Buscar" component={Search} />
-      <Drawer.Screen name='Carrito' component={BuyCart} />
-      <Drawer.Screen name="Mi cuenta" component={LoginUser} />
-      <Drawer.Screen
-        name="RegisterUser"
-        component={RegisterUser}
-        options={{ drawerLabel: () => null, title: null }} 
-      />
+      <Drawer.Screen name="Buscar" component={Search}/>
+      <Drawer.Screen name="Cateogorias" component={Category}/>
+      <Drawer.Screen name="Mis Compras" component={Shopping}/>
+      <Drawer.Screen name="Mis Favoritos" component={Favorite}/>
+      <Drawer.Screen name="Ofertas" component={Offer}/>
+      <Drawer.Screen name="Mi perfil" component={Profile}/>
+      <Drawer.Screen name="Ayuda y Soporte" component={Support}/>
+      
     </Drawer.Navigator>
   );
-};export default HomeScreen;
+}
+
+
+export default HomeScreen;
